requirements.txt
flask==3.0.3
anthropic==0.40.0
python-dotenv==1.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
